+++ lib/generators/ruby_native/templates/add_restored_transaction_id_to_ruby_native_purchase_intents.rb
@@ -1,0 +1,5 @@
+class AddRestoredTransactionIdToRubyNativePurchaseIntents < ActiveRecord::Migration[7.1]
+  def change
+    add_column :ruby_native_purchase_intents, :restored_transaction_id, :string
+  end
+end